refactor: 💡 import path redundancy
