+++ src/components/Accordion/Accordion.stories.tsx
@@ -2,8 +2,8 @@ import { Meta, StoryObj } from '@storybook/react-vite';
 import { Panel } from '@/components/Panel';
 import { Accordion } from '@/components/Accordion';
 import { Spacer } from '@/components/Spacer';
-import { Text } from '@/components/Typography/Text/Text';
-import { Title } from '@/components/Typography/Title/Title';
+import { Text } from '@/components/Typography/Text';
+import { Title } from '@/components/Typography/Title';
 
 const meta: Meta<typeof Accordion> = {
   component: Accordion,
