+++ src/components/ContextMenu/ContextMenu.tsx
@@ -2,8 +2,8 @@ import * as RightMenu from '@radix-ui/react-context-menu';
 import { styled } from 'styled-components';
 import { forwardRef } from 'react';
 import type { HorizontalDirection } from '@/types';
-import { Icon } from '@/components/Icon/Icon';
-import type { IconName } from '@/components/Icon/Icon.types';
+import { Icon } from '@/components/Icon';
+import type { IconName } from '@/components/Icon';
 import { Arrow, GenericMenuItem, GenericMenuPanel } from '@/components/GenericMenu';
 import Popover_Arrow from '@/components/Assets/Icons/Popover-Arrow';
 import { IconWrapper } from '@/components/IconWrapper/IconWrapper';
