+++ src/components/Dropdown/Dropdown.tsx
@@ -4,8 +4,8 @@ import { styled } from 'styled-components';
 import { Arrow, GenericMenuItem, GenericMenuPanel } from '@/components/GenericMenu';
 import Popover_Arrow from '@/components/Assets/Icons/Popover-Arrow';
 import { IconWrapper } from '@/components/IconWrapper';
-import { Icon } from '@/components/Icon/Icon';
-import type { IconName } from '@/components/Icon/Icon.types';
+import { Icon } from '@/components/Icon';
+import type { IconName } from '@/components/Icon';
 import type { HorizontalDirection } from '@/types';
 
 export const Dropdown = (props: DropdownMenu.DropdownMenuProps) => (
