+++ src/components/HoverCard/HoverCard.stories.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { Meta, StoryObj } from '@storybook/react-vite';
 import { Checkbox } from '@/components/Checkbox';
 import { Spacer } from '@/components/Spacer';
-import { Text } from '@/components/Typography/Text/Text';
-import { Title } from '@/components/Typography/Title/Title';
+import { Text } from '@/components/Typography/Text';
+import { Title } from '@/components/Typography/Title';
 import { HoverCard } from '@/components/HoverCard';
 
 interface HoverCardExampleProps {
