+++ src/components/Tabs/Tabs.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Meta, StoryObj } from '@storybook/react-vite';
 import { Spacer } from '@/components/Spacer';
 import { Tabs } from '@/components/Tabs';
-import { Text } from '@/components/Typography/Text/Text';
+import { Text } from '@/components/Typography/Text';
 
 const meta: Meta<typeof Tabs> = {
   component: Tabs,
