+++ src/components/Tooltip/Tooltip.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta, StoryObj } from '@storybook/react-vite';
 import { Tooltip } from '@/components/Tooltip';
-import { Text } from '@/components/Typography/Text/Text';
+import { Text } from '@/components/Typography/Text';
 
 const meta: Meta<typeof Tooltip> = {
   component: Tooltip,
